requirements.txt
argon2-cffi==20.1.0
asn1crypto==0.24.0
async-generator==1.10
attrs==20.3.0
backcall==0.1.0
beautifulsoup4==4.7.1
bleach==3.2.1
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
click==7.1.2
conda==4.6.2
conda-build==3.17.8
cryptography==2.5
cycler==0.10.0
Cython==0.29.21
decorator==4.3.2
defusedxml==0.6.0
entrypoints==0.3
filelock==3.0.10
Flask==1.1.2
glob2==0.6
idna==2.8
importlib-metadata==3.3.0
ipykernel==5.4.2
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.6.2
itsdangerous==1.1.0
jedi==0.13.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kafka-python==2.0.2
kiwisolver==1.3.1
libarchive-c==2.8
lief==0.9.0
MarkupSafe==1.1.0
matplotlib==3.3.3
mistune==0.8.4
mkl-fft==1.0.10
mkl-random==1.0.2
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
notebook==6.1.6
numpy==1.15.4
opencv-python-headless==4.2.0.34
packaging==20.8
pandas==1.1.5
pandocfilters==1.4.3
parso==0.3.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==8.1.0
pkginfo==1.5.0.1
prometheus-client==0.9.0
prompt-toolkit==2.0.8
psutil==5.5.0
ptyprocess==0.6.0
pycocotools==2.0.2
pycosat==0.6.3
pycparser==2.19
Pygments==2.3.1
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.6.8
python-dateutil==2.8.1
pytz==2018.9
PyYAML==3.13
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
requests==2.21.0
ruamel-yaml==0.15.46
Send2Trash==1.5.0
six==1.12.0
soupsieve==1.7.1
terminado==0.9.2
testpath==0.4.4
torch==1.0.1.post2
torchvision==0.2.2
tornado==6.1
tqdm==4.29.1
traitlets==4.3.2
typing-extensions==3.7.4.3
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
zipp==3.4.0
